Cache output wave data
Prior to this change, `fwrite` is called for every samples, which could
cause performance issue.
The audio data is intentionally cached on memory and write the data into
file once some amount is accumulated. In addition, it is controlled to
avoid two or more `write` to multiple files in one packet processing.